Initial restructuring of the repo
