+++ src/routes/index.tsx
@@ -1,20 +1,19 @@
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import {Layout} from '@/components/Layout';
-import {Feed} from './Feed';
-import FormContainer from '@/components/FormContainer';
+import Layout from '@/components/Layout';
+import FeedIndex from './feed/Index';
+import CasesIndex from './cases';
+import CasesNew from './cases/new';
 import {AuthorizationCallback} from '@/hooks/auth';
-import CaseList from '@/components/CaseList';
 
-export function Router() {
+export default function Router() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path='/oauth/cb' element={<AuthorizationCallback />} />
         <Route path='/' element={<Layout />}>
-          <Route index element={<Feed />} />
-          <Route path='/feed' element={<Feed />} />
-          <Route path='/form' element={<FormContainer />} />
-          <Route path='/list' element={<CaseList />} />
+          <Route index element={<FeedIndex />} />
+          <Route path='/cases' element={<CasesIndex />} />
+          <Route path='/cases/new' element={<CasesNew />} />
         </Route>
       </Routes>
     </BrowserRouter>
